Versioning, and unit weights
Added versioning to Session and TimestampedWeight with custom readObject function so they can be saved and retrieved in the future.

Also added here is unit measurement. So it reads bit 14 and applies kg or lb to each timestamped weight. This is then displayed on Live data, Peak, and Repeater screens. Crit force will be next